requirements.txt
openai>=1.30.0
pyyaml>=6.0.1
pytest>=8.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
huggingface-hub>=0.20.0
openenv-core>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
